+++ src/main/java/com/tringles/tutorial/domain/user/User.java
@@ -1,15 +1,13 @@
-package com.tringles.tutorial.domain;
+package com.tringles.tutorial.domain.user;
 
 import com.tringles.tutorial.domain.base.BaseTimeEntity;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
 import javax.persistence.*;
-import java.util.Collection;
 import java.util.Set;
 
 @Data
