+++ src/main/java/com/tringles/tutorial/service/OAuth2UserService.java
@@ -1,4 +1,4 @@
-package com.tringles.tutorial.config.OAuth2;
+package com.tringles.tutorial.service;
 
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
